Convert from the old directive format to the new component format.
This commit has a lot of changes, mostly the conversion from the old
angular directive format to the new 1.5 component format. This makes
maintaining the library a little easier and lightens the code up by
almost half.

Also in this commit is a change to the license. At the request of a
few people, I've gone with an MIT license for this module so they can
make use of ng-epoch for projects within organizations that have
strict licensed software policies. Hopefully this helps you all out
and I'm glad you find this little nugget useful enough to request the
license change.

Other than this, there are no changes to implementation. The tags are
exactly the same as the original directive. The only real changes are
the dependency change on Angular 1.5 and the Epoch version update.